app/requirements.txt
runpod
speedtest-cli
matplotlib
numpy
